refactor uint to uint64 in history tree
Former-commit-id: 523825417d346cc3b61d8e82301e3448e793d641

+++ balloon/history/historytest.go
@@ -55,7 +55,7 @@ func NewFakeCleanTree(frozen storage.Store, hasher hashing.Hasher, l log.Logger)
 	return tree
 }
 
-func NewFakeProof(auditPath []Node, index uint, hasher hashing.Hasher) *Proof {
+func NewFakeProof(auditPath []Node, index uint64, hasher hashing.Hasher) *Proof {
 
 	proof := NewProof(auditPath, index, hasher)
 	proof.leafHasher = fakeLeafHasherF(hasher)
